+++ test/browserbehavior/DefaultBrowserBehavior.test.ts
@@ -253,11 +253,17 @@ describe('DefaultBrowserBehavior', () => {
   });
 
   describe('requiresSimulcastMunging', () => {
-    it('can determine if simulcast requires munging', async () => {
+    it('can determine if simulcast requires munging', () => {
       setUserAgent(CHROME_USERAGENT);
       expect(new DefaultBrowserBehavior().requiresSimulcastMunging()).to.eq(false);
       setUserAgent(SAFARI_USERAGENT);
       expect(new DefaultBrowserBehavior().requiresSimulcastMunging()).to.eq(true);
+    });
+  });
+
+  describe('supportsCanvasCapturedStreamPlayback', () => {
+    it('can determine if video processing is supported', () => {
+      expect(new DefaultBrowserBehavior().supportsCanvasCapturedStreamPlayback()).to.eq(true);
     });
   });
 });