+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Samsung Internet browser for Android as a supported browser.
+- Add `DefaultVideoTransformDevice` to implement `VideoTransformDevice`.
+ `VideoFrameProcessor`, `VideoFrameProcessorPipeline` and `VideoFrameBuffer` interfaces
+ are added to support `DefaultVideoTransformDevice` and allow processing steps to be applied to device.
+ The method `chooseVideoInputDevice` in `DefaultDeviceController` can handle `VideoTransformDevice` now.
 
 ### Changed